c: Make integer casts a little less "warny"
On LLP64 systems (e.g. Windows) 64-bit pointers are cast to 32-bit
integers, which tends to set off warnings. Losing those upper bits is
perfectly fine for this application, and compilers generally don't care
about the data loss when casting from integer to integer.

+++ c/example.c
@@ -47,9 +47,9 @@ main(int argc, char **argv)
     } else {
         void *p = malloc(4UL * 1024 * 1024);
         *seed ^= hash32(time(0));             /* Current time */
-        *seed ^= hash32((unsigned long)main); /* ASLR entopy */
-        *seed ^= hash32((unsigned long)seed); /* Stack gap entropy */
-        *seed ^= hash32((unsigned long)p);    /* Allocator entropy */
+        *seed ^= hash32((ptrdiff_t)main);     /* ASLR entopy */
+        *seed ^= hash32((ptrdiff_t)seed);     /* Stack gap entropy */
+        *seed ^= hash32((ptrdiff_t)p);        /* Allocator entropy */
         free(p);
     }
